### 2023-01-11 #5069 @editorialbot
[PRE REVIEW]: excitingtools: An exciting Workflow Tool
**Submitting author:** @AlexBuccheri (<a href="http://orcid.org/0000-0001-5983-8631">Alexander Buccheri</a>)
**Repository:** https://github.com/exciting/excitingtools
**Branch with paper.md** (empty if default branch):
**Version:** v1.1.0
