### 2023-01-21 f4c82e7
JOSS Paper Bibtext DOIs (#5)
* JOSS paper bibtex DOIs
