+++ paper/paper.bib
@@ -8,7 +8,8 @@
   number={36},
   pages={363202},
   year={2014},
-  publisher={IOP Publishing}
+  publisher={IOP Publishing},
+  doi={10.1088/0953-8984/26/36/363202}
 }
 
 @article{gulans:2018,
@@ -19,7 +20,8 @@
   number={16},
   pages={161105},
   year={2018},
-  publisher={APS}
+  publisher={APS},
+  doi={10.1103/physrevb.97.161105}
 }
 
 @article{nabok:2016,
@@ -30,7 +32,8 @@
   number={3},
   pages={035118},
   year={2016},
-  publisher={APS}
+  publisher={APS},
+  doi={10.1103/PhysRevB.94.035118}
 }
 
 @article{lejaeghere:2016,
@@ -42,7 +45,8 @@
   number={6280},
   pages={aad3000},
   year={2016},
-  publisher={American Association for the Advancement of Science}
+  publisher={American Association for the Advancement of Science},
+  doi={10.1126/science.aad3000}
 }
 
 @article{larsen:2017,
@@ -54,7 +58,8 @@
   number={27},
   pages={273002},
   year={2017},
-  publisher={IOP Publishing}
+  publisher={IOP Publishing},
+  doi={10.1088/1361-648X/aa680e}
 }
 
 @article{mathew:2017,
@@ -65,7 +70,8 @@
   volume={139},
   pages={140--152},
   year={2017},
-  publisher={Elsevier}
+  publisher={Elsevier},
+  doi={10.1016/j.commatsci.2017.07.030}
 }
 
 @misc{Ganose:2022,
@@ -85,7 +91,8 @@
   volume={199},
   pages={110731},
   year={2021},
-  publisher={Elsevier}
+  publisher={Elsevier},
+  doi={10.1016/j.commatsci.2021.110731}
 }
 
 @article{gavini:2022,
@@ -93,7 +100,8 @@
   author={Gavini, Vikram and Baroni, Stefano and Blum, Volker and Bowler, David R and Buccheri, Alexander and
   Chelikowsky, James R and Das, Sambit and Dawson, William and Delugas, Pietro and Dogan, Mehmet and others},
   journal={arXiv preprint arXiv:2209.12747},
-  year={2022}
+  year={2022},
+  doi={10.48550/arXiv.2209.12747}
 }
 
 @article{jain:2013,
@@ -105,7 +113,8 @@
   number={1},
   pages={011002},
   year={2013},
-  publisher={American Institute of PhysicsAIP}
+  publisher={American Institute of PhysicsAIP},
+  doi={10.1063/1.4812323}
 }
 
 @article{curtarolo:2012,
@@ -116,7 +125,8 @@
   volume={58},
   pages={218--226},
   year={2012},
-  publisher={Elsevier}
+  publisher={Elsevier},
+  doi={10.1016/j.commatsci.2012.02.005}
 }
 
 @article{draxl:2019,
@@ -127,5 +137,6 @@
   number={3},
   pages={036001},
   year={2019},
-  publisher={IOP Publishing}
+  publisher={IOP Publishing},
+  doi={10.1088/2515-7639/ab13bb}
 }
