### 2023-01-21 #5069 @AlexBuccheri
Hi @kyleniemeyer thanks very much for starting the submission process.

With regards to the DOI issue, could you clarify what the editorial bot is flagging up, is it missing the 'doi' field in the bibtex? If so, I've added DOIs and pushed back to the `excitingtools` repo.

I’ve had a look through the reviewer list. Anyone with python as their preferred language, and (ideally) workflows or electronic structure as a topic of interest would be appropriate. _cmbiwer_ or _vmatos_ could be appropriate, for example.
