### 2023-02-20 #5148 @fjmartinmartinez
* In readme.md line 105 it seems to be a to-do task. Is this addressed? Should I rise this as an issue in the code?

* In the paper it is stated that "excitingtools (plus dependencies) is significantly more lightweight than the NOMAD package, approximately two orders of magnitude smaller". Maybe some comment on what do you mean by "lightweight" would be good. I am assuming it means disk space or RAM memory?

* Suggestion for improvement (just as ideas):
    * Use of conda/mamba as alternative for installation, if possible.
    * In addition, the workability of the tool could be improved by defining a workspace for vscode
