### 2023-02-23 b37a323
Remove TODO and improve formatting
