+++ README.md
@@ -1,27 +1,27 @@
 # excitingtools
-<span style="font-family:american typewriter; font-size:1em;">**excitingtools**</span> is a collection of 
+
+<span style="font-family:american typewriter; font-size:1em;">**excitingtools**</span> is a collection of
 modules to facilitate the generation of <span style="font-family:american typewriter; font-size:1em;">**exciting**</span>
-inputs and the post-processing of <span style="font-family:american typewriter; font-size:1em;">**exciting**</span> outputs. 
+inputs and the post-processing of <span style="font-family:american typewriter; font-size:1em;">**exciting**</span> outputs.
 
 <span style="font-family:american typewriter; font-size:1em;">**excitingtools**</span> currently provides functionality for:
 
-* Generation of the <span style="font-family:american typewriter; font-size:1em;">**exciting**</span> input XML file 
+* Generation of the <span style="font-family:american typewriter; font-size:1em;">**exciting**</span> input XML file
   using Python classes:
-  - Currently supported for `groundstate`, `structure` and `BSE`
-
+  * Currently supported for `groundstate`, `structure` and `BSE`
 
 * Parsing of <span style="font-family:american typewriter; font-size:1em;">**exciting**</span> outputs into Python dictionaries
 
-
 * High-level class API for interacting with results:
-  - Currently implemented for eigenvalues, band structure and DOS (without SO coupling)
+  * Currently implemented for eigenvalues, band structure and DOS (without SO coupling)
 
-making it is possible to define a calculation, run it, and parse the relevant outputs all from within Python. 
+making it is possible to define a calculation, run it, and parse the relevant outputs all from within Python.
 
 <span style="font-family:american typewriter; font-size:1em;">**excitingtools**</span> is used by, or in conjunction with:
+
 * <span style="font-family:american typewriter; font-size:1em;">**exciting's**</span> regression-testing framework
   * Parsing of output data
-* <span style="font-family:american typewriter; font-size:1em;">**exciting's**</span> Jupyter notebook tutorials 
+* <span style="font-family:american typewriter; font-size:1em;">**exciting's**</span> Jupyter notebook tutorials
   * Data handling
 * [Atomic Simulation Environment (ASE)](https://wiki.fysik.dtu.dk/ase/)
   * Input and output handling in ASE's <span style="font-family:american typewriter; font-size:1em;">**exciting**</span> calculator
@@ -29,7 +29,8 @@ making it is possible to define a calculation, run it, and parse the relevant ou
   * For the development of complex, automated <span style="font-family:american typewriter; font-size:1em;">**exciting**</span> workflows  
 
 ## Installation
-If one wishes to import <span style="font-family:american typewriter; font-size:1em;">**excitingtools**</span> in their own scripts, it can be installed from this project's root directory 
+
+If one wishes to import <span style="font-family:american typewriter; font-size:1em;">**excitingtools**</span> in their own scripts, it can be installed from this project's root directory
 (`$EXCITING_ROOT/tools/exciting_tools`) with:
 
 ```bash
@@ -43,12 +44,14 @@ pip install excitingtools
 ```
 
 ## External Package Dependencies
-If a new external dependency is introduced to the package, this also requires adding to `setup.py` such that pip is aware 
+
+If a new external dependency is introduced to the package, this also requires adding to `setup.py` such that pip is aware
 of the new dependency.
 
-## Basic File Structure 
+## Basic File Structure
+
 In general, modules should begin with a docstring giving an overview of the module's purpose. External python
[…]
