### 2023-02-23 #5148 @AlexBuccheri
@fjmartinmartinez thanks for the review. I'll verbally address the points here and an open a PR from this issue to make the proposed changes:

> In readme.md line 105 it seems to be a to-do task. Is this addressed? Should I rise this as an issue in the code?

_excitingtools_ has been calved off from a bigger repository and the associated issue was being tracked in that repo's issue tracker. As it's not meaningful here, and  pdoc does a perfectly good job at generating documentation, I'll remove it.

- [ ] Remove issue 105

> In the paper it is stated that "excitingtools (plus dependencies) is significantly more lightweight than the NOMAD package, approximately two orders of magnitude smaller". Maybe some comment on what do you mean by "lightweight" would be good. I am assuming it means disk space or RAM memory?

I was referring to package size, but after making some effort to double-check this, I believe I am just factually incorrect. I can't straightforwardly verify the larger packages as `pip install nomad-lab[parsing]`  and `pip install
nomad-lab[all]`  fail to install on both my mac M1 ( Python 3.9.16) and Debian systems (Python 3.7.3) for various reasons.
However, the minimal nomad-lab package, which provides the [exciting parsers](https://github.com/nomad-coe/nomad-parser-exciting) is ~ 2.4 times larger than excitingtools  (quantified by installing the respective packages in venvs and looking at the size diff). There's a noticeable difference in the number of package dependencies (see below), but in general, the statement is ambiguous and not meaningful, so I think we can remove it.

- [ ]  Remove the statement: "_excitingtools_ (plus dependencies) is significantly more lightweight than the NOMAD package, approximately two orders of magnitude smaller."

excitingtools dependencies:
`contourpy-1.0.7 cycler-0.11.0 fonttools-4.38.0 importlib-resources-5.12.0 kiwisolver-1.4.4 matplotlib-3.7.0 numpy-1.24.2 packaging-23.0 pillow-9.4.0 pyparsing-3.0.9 python-dateutil-2.8.2 six-1.16.0 wheel-0.38.4 zipp-3.14.0`

NOMAD dependencies:
`Pint-0.17 aniso8601-7.0.0 anyio-3.6.2 ase-3.19.0 cachetools-4.2.4 certifi-2022.12.7 charset-normalizer-2.0.12 click-7.1.2 contourpy-1.0.7 cycler-0.11.0 cython-0.29.33 docstring-parser-0.12 ecdsa-0.18.0 elasticsearch-7.17.9 elasticsearch-dsl-7.4.0 excitingparser-1.0 fastentrypoints-0.12 fonttools-4.38.0 future-0.18.2 h11-0.12.0 httpcore-0.14.7 httpx-0.22.0 idna-3.4 importlib-resources-5.12.0 jmespath-0.10.0 kiwisolver-1.4.4 matplotlib-3.7.0 memoization-0.4.0 nomad-lab-1.0.8 nptyping-1.4.4 numpy-1.21.2 orjson-3.6.0 packaging-23.0 pillow-9.4.0 pyasn1-0.4.8 pydantic-1.8.2 pyparsing-3.0.9 python-dateutil-2.8.2 python-jose-3.3.0 python-keycloak-0.26.1 pytz-2021.1 pyyaml-6.0 requests-2.26.0 rfc3986-1.5.0 rsa-4.9 scipy-1.10.1 six-1.16.0 sniffio-1.3.0 typing-extensions-4.5.0 typish-1.9.3 urllib3-1.26.14 wheel-0.38.4 zipp-3.14.0`

## Suggestions For Improvements

> Use of conda/mamba as alternative for installation, if possible.

I agree with this, but would like to delay this until the summer as I expect the minor version to bump several times in the next 6 months (and would like to reduce the overhead of deploying releases).

> In addition, the workability of the tool could be improved by defining a workspace for vscode

 I interpret this to mean providing `settings.json`, `tasks.json`, and ` launch.json` files. I can add some basic settings that specify formatting, linting and tests. @fjmartinmartinez did you have something more in mind? Additionally I'd be interested to see any snippets that install package requirements into a venv, then ensure all new VS code terminals use this venv. I must admit I use Pycharm for all of my Python development and find it far more pleasurable to work in (but default to VS code for Fortran).

- [ ] Add VS code workspace files to repo
