### 2023-02-23 5f2679a
Remove ambiguous line refering to relative NOMAD package size
