### 2023-03-04 #5148 @hmacdope
The package is neat, functional and well maintained. Well done @AlexBuccheri!

I would also like to suggest a conda release, this will help you get more users, but is not an absolute requirement.

- [ ] I would however as mentioned in the paper love to see an example of how to integrate with `JobFlow`, `atomate` or `ASR`, just so users can see the tool in action and how to perform the proposed integration with higher level workflow managers.

For the paper:

I would love a few more citations for some of the more general statements eg
-  "Moreover, machine learning models that predict material properties would greatly benefit from the availability of
 higher-fidelity data sets for a range of systems."
- "the choice of the LAPW basis and systematic convergence of calculations, even at the ground state level of theory, is more involved than with plane wave or Gaussian type orbital (GTO) basis sets. "
- A few others

Other than that looks great!
