### 2023-03-08 #5148 @fjmartinmartinez
@AlexBuccheri sorry, I did not follow up on your response. All looks great to me. As @hmacdope said, the package is neat. What I meant by the workspace is just this: https://code.visualstudio.com/docs/editor/workspaces. Nothing else in mind from my side, and it was just a suggestion. Other than that, well done!
