### 2023-03-21 #5148 @AlexBuccheri
@fjmartinmartinez and @hmacdope thank you both very much for the very positive, constructive reviews.

To address @hmacdope's queries, I'll reply here and put any required changes in the linked [MR](https://github.com/exciting/excitingtools/pull/6)

**Example Integration with Jobflow or ASR**

To add a few points for context:

a) Workflow development is very much a work-in-progress (a PhD project)

- Usage with Jobflow depends on object serialisation, and unfortunately that library does not support pickling, which is apparently not suitably transferable. Fabian is still finalising serialisation methods for classes in excitingtools.

b) The workflow package is a separate repository, such that responsibilities are clearly separated. Until it reaches a stable version, it will stay on the group’s private Gitlab instance. For now, it will provide single responsibility functions to build workflows and contain the workflows themselves. Although this may change in the future.

c) ASR version 2 is undergoing a rewrite, and perhaps will be released under a different name. Its API is not stable and several features are still missing (being able to treat convergence, for example).

**Jobflow Example**

 A workflow is defined as a series of single-responsibility function calls:

 * Single-responsibility functions take a calculation (or a result)

 * Result-handling with Jobflow is included with function decorators,
   which allow lazy evaluation, meta-data storage, etc.

 * A calculation is a specialisation of the CalculationIO base class
   - This is under active development with ASE
   - The specialisation methods are injected from excitingtools.

```python
# Simple convergence workflow, composed of steps defined by us
# Composed and executed by Joblow

""" Execute convergence. """
# read input from input file:
specified_input = read_input(Path("input.yml").absolute())

convergence_criteria = get_convergence_criteria(specified_input)

# setup jobs:
setup_job = setup_exciting_calculation(specified_input)
convergence_job = converge_ngridk(setup_job.output, convergence_criteria, [])

# run workflow:
responses = run_locally(Flow([setup_job, convergence_job]),
store=JobStore(MemoryStore()))
print(responses)
```
Functions responsible for performing workflow "steps".

In this example, we're performing and evaluating convergence of k-sampling - This could probably be more generic (and I'll catch that in code review).

```python
from typing import Callable

from jobflow import job, Response

from excitingworkflow.src.convergence_criteria.exciting_spectra_convergence_criteria
import set_ngridk_in_input
from excitingworkflow.src.workflows.workflow_results import ConvergenceResult

def converge_recursively_step(calculation, input_value,set_value_in_input: Callable) -> dict:
     """ Perform a single convergence step.
     :param calculation: define all calculation parameters, how to run and parse results
     :param input_value: the current input_value
     :param set_value_in_input: how to put the input value in the calculation
     :returns: a dictionary with the results from the current run
     """
     set_value_in_input(input_value, calculation)
     subprocess_result = calculation.run()

     if not subprocess_result.success:
         raise RuntimeError(f"exciting run was not successful!")
     return calculation.parse_output()

@job
def converge_ngridk(calculation, convergence_criteria, results: list):
     """ Converge ngridk.
     :param calculation: calculation with all the other parameters, how to run and get results
     :param convergence_criteria: defines the convergence, how to check, which steps to take and when to end
     :param results: capture the convergence results
     :returns: the results.
     """
     input_value, is_first, is_last = convergence_criteria.get_current_input_value()
     output = converge_recursively_step(calculation, input_value, set_ngridk_in_input)

     if is_first:
         converged = False
         result = ConvergenceResult(input_value, calculation.directory, output, converged, early_exit=False)
         results = [result]

     # check for convergence using the convergence criteria evaluate function
         converged, early_exit = convergence_criteria.evaluate(output, results[-1].result)
         result = ConvergenceResult(input_value, calculation.directory, output, converged, early_exit)
         results.append(result)

     if converged or is_last:
         return results

     # else append the workflow DYNAMICALLY with a new workflow step
     return Response(output=results, replace=converge_ngridk(calculation, convergence_criteria, results))
```

```python
# An exciting calculation is defined with an IO calculator.
# Below is the abstract class for an IO calculator.
# A specific implementation utilitises the methods provided by excitingtools.

 class CalculationIO(abc.ABC):
     """Abstract base class for a calculation that is performed
     by writing input file/s and parsing the result from a file.

     An IO calculation is expected to have:
         * A name,
         * A working (run) directory,
         * A method to write all input files required to run the calculation,
         * A method to run the calculation,
         * A parser for the outputs of interest.
     """
     path_type = Union[str, Path]

     def __init__(self, name: str, directory: path_type):
         self.name = name
         self.directory = Path(directory)

     def set_name(self, name: str):
         """ Set name to given argument. """
         self.name = name

     @abc.abstractmethod
     def write_inputs(self) -> None:
         """ Make run dir and write all input files required for calculation.
         """
         self.directory.mkdir(parents=True, exist_ok=True)
         ...

     @abc.abstractmethod
     def run(self) -> SubprocessRunResults:
         """ Run the calculation.
         :return: Subprocess result instance.
         """
         ...

     @abc.abstractmethod
     def parse_output(self, *args) -> Union[dict, FileNotFoundError]:
         """ Parse one or more output files for calculation.
         :return: Dictionary of results.
         """
         ...
```

**With regards to using ASR**

The ASE team has approved the integration of excitingtools into ASE. This is part of ASE's program to standardise its calculator class, and migrate implementation to client codes. This means that the ASE exciting calculator wraps excitingtools functions and uses them to create inputs, read outputs and run calculations. Atomic Simulation Recipes (ASR) integrates with ASE calculators so excitingtools is never _explicitly_ called in ASR workflows using exciting.

Collaboration with the ASR team to create exciting workflows is currently under active development. The ASR framework is undergoing massive changes, which need to stabilise before we can start utilising it in production. We can however run simple workflows, such as this adapted from ASR's tutorials:

```python
import ase
import asr

@asr.instruction('asr.tutorial')
@asr.argument('crystal_structure', type=str)
@asr.argument('element', type=str)
def energy_per_atom(element: str, crystal_structure: str) -> float:
    """Calculate the total energy per atom of material"""
    atoms = bulk(element, crystalstructure=crystal_structure)
    cls = ase.get_calculator_class('exciting')
    atoms.calc = cls()
    energy = atoms.get_potential_energy()
    num_atoms = len(atoms)
    return round(energy / num_atoms, 6)

@asr.instruction('asr.tutorial')
@asr.argument('element', type=str)
def main(element: str) -> str:
    """Calculate lowest energy crystal structure for an element."""
    crystal_structures = [
        'sc', 'fcc', 'bcc', 'diamond',
    ]
    energies = []
    for crystal_structure in crystal_structures:
        en = energy(element, crystal_structure)
        energies.append(en)
    lowest_energy_crystal_structure = min(
        zip(crystal_structures, energies),
        key=lambda item: item[1]
    )[0]

    return lowest_energy_crystal_structure
```
