### 2023-03-21 #5148 @AlexBuccheri
> Conda

As both @fjmartinmartinez  and @hmacdope recommend a Conda release,  we can facilitate this

> ML Training Data References

I'll introduce these citations regarding the need for training data (this broadly applies to all fields):

- https://link.springer.com/referenceworkentry/10.1007/978-3-319-44677-6_104
- https://doi.org/10.1093/bib/bbab391
- https://www.nature.com/articles/s42256-021-00319-w
