### 2023-03-24 #5148 @hmacdope
@AlexBuccheri thanks for your thoughtful and detailed replies.

- RE jobflow and other workflow automation tools, if you could make a note in the README or docs somewhere that you are actively developing the interfaces to workflow codes that would be awesome.  If this already exists then ignore.
- RE  LAPW basis set complexity, your very detailed answer is very helpful. Happy to move forward.

Other than that I am totally satisfied, the wait on full conda-package deployment can be a little while, but I will be happy to accept when you link a PR adding `excitingtools` against `conda-forge/staged-recipies`. Let me know if you think this is unreasonable. : )
