### 2023-03-26 #5148 @AlexBuccheri
[MR 6](https://github.com/exciting/excitingtools/pull/6) has been merged, implementing the changes requested in this review. That should resolve all remaining threads.
