### 2023-04-04 #5148 @AlexBuccheri
@richardjgowers I believe this covers the TODOs:

excitingtools [Version 1.3.0](https://github.com/exciting/excitingtools/releases/tag/1.3.0)

[![DOI](https://zenodo.org/badge/564833158.svg)](https://zenodo.org/badge/latestdoi/564833158)
