+++ paper/paper.bib
@@ -147,7 +147,8 @@
   journal={Handbook of Materials Modeling: Methods: Theory and Modeling},
   pages={49--73},
   year={2020},
-  publisher={Springer}
+  publisher={Springer},
+  doi = {10.1007/978-3-319-44677-6_104}
 }
 
 @article{toniato:2021,
@@ -158,7 +159,8 @@
   number={6},
   pages={485--494},
   year={2021},
-  publisher={Nature Publishing Group UK London}
+  publisher={Nature Publishing Group UK London},
+  doi = {10.1038/s42256-021-00319-w}
 }
 
 @article{dong:2022,
@@ -169,7 +171,8 @@
   number={1},
   pages={bbab391},
   year={2022},
-  publisher={Oxford University Press}
+  publisher={Oxford University Press},
+  doi = {10.1093/bib/bbab391}
 }
 
 @article{blaha:2020,
@@ -180,7 +183,8 @@
   number={7},
   pages={074101},
   year={2020},
-  publisher={AIP Publishing LLC}
+  publisher={AIP Publishing LLC},
+  doi = {10.1063/1.5143061}
 }
 
 @article{andersen:1975,
@@ -191,5 +195,6 @@
   number={8},
   pages={3060},
   year={1975},
-  publisher={APS}
+  publisher={APS},
+  doi = {10.1103/physrevb.12.3060}
 }
