### 2023-04-06 #5148 @AlexBuccheri
@richardjgowers Sorry, I was only looking INVALID DOIs. Should be resolved now.
