### 2023-04-06 d587312
Do not pin setuptools_scm to old versions. The first version available on conda-forge is 3.5.0
