### 2023-05-03 #5148 @AlexBuccheri
Hi @kyleniemeyer ah, the difference was the PI. She is not in the` CONTRIBUTING.md` having not worked on the code, but was part of the paper. I've amended the information on Zenodo to include her.
