### 2023-05-03 #5148 @kyleniemeyer
@AlexBuccheri I've been waiting for the PDF itself to appear before closing the issue, but yes otherwise this looks good to go. Congratulations on your article's publication in JOSS!
